root-service-manager tests
